Fullførte refaktorering av logikk- og rendering-håndtering
